+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `description` and `state` properties to the `RadioGroup` component
 
+### Fixed
+
+- Fixed a margin bottom issue with the inline alert title
+
 ## 7.11.1 - 2026-01-07
 
 ### Fixed